LayoutTests/platform/gtk/accessibility/canvas-description-and-role-expected.txt

This test makes sure that a canvas with and without fallback content each has the right role and description.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axContainer.childrenCount is 2
Canvas 1 description: AXDescription: Canvas label
Canvas 1 role: AXRole: canvas
Canvas 2 description: AXDescription: 
Canvas 2 role: AXRole: canvas
PASS successfullyParsed is true

TEST COMPLETE
